+++ src/core/main.ts
@@ -6,7 +6,7 @@ import {
   propertyAccessExpressionCreatorWithTarget,
   ValueOf,
   zodIdentifier,
-} from "../utils/helpers";
+} from "../utils/ast";
 
 import {
   zodArrayMembers,
